modbus_client/requirements.txt
pymodbus
python-dotenv
pymongo
requests>=2.0
